+++ src/scrappie_structures.h
@@ -1,8 +1,6 @@
 #ifndef SCRAPPIE_STRUCTURES_H
 #define SCRAPPIE_STRUCTURES_H
 
-#include <stdlib.h>
-
 typedef struct {
         double start;
         float length;
@@ -11,12 +9,12 @@ typedef struct {
 } event_t;
 
 typedef struct {
-        size_t n, start, end;
+        unsigned int n, start, end;
         event_t * event;
 } event_table;
 
 typedef struct {
-        size_t n, start, end;
+        unsigned int n, start, end;
         float * raw;
 } raw_table;
